fastapi_resnet/requirements.txt
fastapi==0.88.0
Pillow==9.3.0
pytest==7.1.1
uvicorn==0.20.0
python-multipart==0.0.5
